RE #INTRD-11258 recalculate advances when update/cancel invoices

Approved-by: opencellsoft

+++ opencell-admin/ejbs/src/main/java/org/meveo/service/billing/impl/LinkedInvoiceService.java
@@ -21,6 +21,7 @@ import java.util.List;
 
 import javax.ejb.Stateless;
 
+import org.apache.commons.collections.CollectionUtils;
 import org.meveo.model.billing.InvoiceTypeEnum;
 import org.meveo.model.billing.LinkedInvoice;
 import org.meveo.service.base.PersistenceService;
@@ -34,12 +35,15 @@ import org.meveo.service.base.PersistenceService;
 @Stateless
 public class LinkedInvoiceService extends PersistenceService<LinkedInvoice> {
     
-    public void deleteByIdInvoiceAndLinkedInvoice(Long invoiceId, List<Long> linkedInvoice) {
-      getEntityManager().createNamedQuery("LinkedInvoice.deleteByIdInvoiceAndLinkedInvoice")
-                            .setParameter("invoiceId", invoiceId)
-                            .setParameter("linkedInvoiceId", linkedInvoice)
-                            .executeUpdate();
-    }
+	public void deleteByIdInvoiceAndLinkedInvoice(Long invoiceId, List<Long> linkedInvoice) {
+		if (CollectionUtils.isEmpty(linkedInvoice)) {
+			return;
+		}
+		getEntityManager().createNamedQuery("LinkedInvoice.deleteByIdInvoiceAndLinkedInvoice")
+				.setParameter("invoiceId", invoiceId)
+				.setParameter("linkedInvoiceId", linkedInvoice)
+				.executeUpdate();
+	}
 
 
 	public void deleteByInvoiceIdAndType(Long invoiceId, InvoiceTypeEnum type) {
